backend/requirements.txt
fastapi==0.115.11
uvicorn[standard]==0.34.0
pydantic-settings==2.8.1
python-multipart==0.0.20
sentence-transformers==3.4.1
scikit-learn==1.6.1
numpy==2.2.3
pdfplumber==0.11.5
python-docx==1.1.2
google-api-python-client==2.164.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
